added servo support

+++ src/main/monitor.cpp
@@ -11,7 +11,7 @@ void monitorSetup(){
     display.setBrightness(7);
 }
 
-void showOneDecimalPrecision(int num){
+void showOneDecimalPrecision(float num){
     int roundedToOneDecimal = (int) (num * 10 + 0.5);
     display.showNumberDecEx(roundedToOneDecimal, 0b00010000);
 }
