+++ src/main/monitor.h
@@ -3,5 +3,5 @@
 
     //Prototype for helper_function found in HelperFunctions.cpp
     void monitorSetup( void );
-    void showOneDecimalPrecision( int num );
+    void showOneDecimalPrecision( float num );
 #endif
